refactor(step2): Step 별 패키지 분리
