+++ src/main/kotlin/study/step2/StringCalculator.kt
@@ -1,4 +1,4 @@
-package study
+package study.step2
 
 import java.lang.IllegalArgumentException
